requirements.txt
tokenizers==0.21.1 
tqdm==4.67.1 
transformers==4.52.4
faiss-cpu==1.11.0 
sentence-transformers==4.1.0
numpy==2.2.4
flask-cors==6.0.0
google-genai
dotenv
